requirements.txt
python-telegram-bot==13.0
aiohttp==3.7.3
torch==0.4.1
torchvision==0.2.2
opencv-python==3.4.2.16
pillow==5.4.0
